Fixes #1 - New RPi firmware breaks Resizer behaviour
I turned on debugging with -DOMX_IMAGE_DEBUG and saw it got stuck at:

OMX EVENT WAITING FOR GROUP: 1

On the old firmware the message that allowed it to get past that wait point needed to be:

>> RECV OMX EVENT: COMP: 3 CLASS: 0 S1: 4 (0x00000004) S2: 61 (0x0000003d) S3: 1 (0x00000001)

but on the new firmware I was instead getting:
>> RECV OMX EVENT: COMP: 3 CLASS: 0 S1: 4 (0x00000004) S2: 61 (0x0000003d) S3: 17 (0x00000011)

That's an OMX_EventBufferFlag event from the Resizer component when it's finished. S3 is nData2 from the EventHandler, and as per https://www.khronos.org/registry/omxil/specs/OpenMAX_IL_1_1_2_Specification.pdf section 3.1.1.4.5:

"The nData2 field of EventHandler specifies the unaltered nFlags field containing the end-of-stream (EOS) flag."

the nFlags set in the code was:

 header->nFlags = OMX_BUFFERFLAG_EOS | OMX_BUFFERFLAG_ENDOFFRAME;

(which is 0x1 | 0x10 = 0x11)

For whatever reason in the old firmware it was returning nData2 as if only OMX_BUFFERFLAG_EOS was set (returning 0x01), and that's what I was forced to check for before. Guess it's got a little closer to the OMX spec now with this update (returning 0x11).

Since I'm now checking for just 0x11 instead of 0x01 in nData2 this will no longer work with the old firmware.